requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.1
requests==2.31.0
pydantic==1.10.13
supabase==2.3.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9 